Change method of installing cisagov/domain-scan
Instead of cloning the repo and acquiring history we do not need we
instead download a tarball of the latest state of the repository on
GitHub and extract it to the desired location. This mirrors how we
install cisagov/domain-scan in the cisagov/scanner image and nets us
~300MiB of saved space between installing curl instead of git and
elminating the unneeded git history (~110MiB are just the git history).